% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_sig_reg.r
\name{call_sig_reg}
\alias{call_sig_reg}
\title{call_sig_reg}
\usage{
call_sig_reg(res, nr, threshold = 3.5, win = 10)
}
\arguments{
\item{res}{An object returned from the sharpr2 function.}

\item{nr}{An integer indicating which region in re to be plotted.}

\item{threshold}{The FWER to control. The default is 3.5.}

\item{win}{A window size for removing sporadic significant regions. If a significant consecutive region is small than win, it will be treated as false signals. The default is 10.}

}
\value{
sig_reg: identified significant regions.

motif: predicted 20bp motifs
}
\description{
Given an object returned from the sharpr2 function, this function calls significant regions that contain motifs.
}
\examples{
# call_sig_reg(re,1)
}
\keyword{ATAC-STAR}
\keyword{sharpr2}
