% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_tile_reg.r
\name{call_tile_reg}
\alias{call_tile_reg}
\title{call_tile_reg}
\usage{
call_tile_reg(data)
}
\arguments{
\item{data}{A data.frame for an ATAC-STAR dataset for one chromosome. The data.frame must contain four columns: 'start', 'end', 'PLASMID', 'RNA', and is sorted by 'start'.}
}
\value{
tile_reg: A list containing the row ids in the data for each tile region.

size the: number of reads in each tile region.

num_r: the total number of tile regions.
}
\description{
For an ATAC-STAR dataset on a given chromosome, this function calls tile regions (the regions covered by at least one read).
}
\examples{
# call_tile_reg(data)
}
\keyword{ATAC-STAR}
\keyword{sharpr2}
