% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reg.r
\name{find_reg}
\alias{find_reg}
\title{find_reg}
\usage{
find_reg(re,pos)
}
\arguments{
\item{re}{An object obtained from sharpr2.}
\item{pos}{A position for which the tile region is searched.}
}
\value{
ind: the index of the tile region in the object from sharpr2. If no such tile region is found, NA is returned.
}
\description{
Given an object from sharpr2 and a position, this function finds the tile region containing the position.
}
\examples{
# find_reg(re,100000)
}
\keyword{ATAC-STAR}
\keyword{sharpr2}
