\name{hidra_ex}
\alias{hidra_ex}
\docType{data}
\title{
An example dataset including a region of one chromosome from a HiDRA library  
}
\description{
This is an example dataset containing 10000 fragments with four columns 'start', 'end', 'PLASMID', 'RNA'. 
}
\usage{data(hidra_ex)}

\format{
  The format is a data.frame with the columns:
	start: the start position of the fragment.
	end: the end position of the fragment.
	PLASMID: the count of PLASMID for this fragment.
	RNA: the count of RNA for this fragment.
}

\examples{
data(hidra_ex)
}
\keyword{datasets}
