% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_sig_reg.r
\name{call_fdr}
\alias{call_fdr}
\title{call_fdr}
\usage{
call_fdr(whole_re, thres_tr = 10, method = 'BH')
}
\arguments{
\item{whole_re}{A list of objects of class sharpr2 for each chromosome.}

\item{thres_tr}{The threshold for the size of tiled reigons used for calculate FDR-adjusted p-values. The default value is 10.}

\item{method}{The method for calculating FDR-adjusted p-values. See the function 'p.adjust' for more details abount the method. The default is 'BH'.}

}
\value{
gfdr: a result table of FDR-adjusted p-values.

}
\description{
Calculate FDR-adjusted p-values.
}
\examples{
# call_fdr(whole_re)
}
\keyword{ATAC-STAR}
\keyword{sharpr2}
