% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_tile_reg.r
\name{call_tile_reg}
\alias{call_tile_reg}
\title{call_tile_reg}
\usage{
call_tile_reg(data)
}
\arguments{
\item{data}{A data.frame for a HiDRA dataset for one chromosome. The data.frame must contain four columns: 'start', 'end', 'PLASMID', 'RNA', and is sorted by 'start'.}
}
\value{
tile_reg: A list containing the row ids in the data for each tiled region.

size: The number of reads in each tiled region.

num_r: The total number of tiled regions.
}
\description{
For a HiDRA dataset on a given chromosome, this function calls tile regions (the regions covered by at least one read).
}
\examples{
data(hidra_ex)
tiled <- call_tile_reg(hidra_ex)
}
\keyword{HiDRA}
\keyword{sharpr2}
