\name{sharpr2-package}
\alias{sharpr2-package}
\alias{sharpr2}
\docType{package}
\title{
Estimating regularoty scores and identifying HDE for HiDRA data
}
\description{
The package develops an algorithm for handling HiDRA datasets by extending MPRA-SHARPR. 
}
\details{
\tabular{ll}{
Package: \tab sharpr2\cr
Type: \tab Package\cr
Version: \tab 0.1.0.0000 \cr
Date: \tab 2017-09-17\cr
License: \tab GPL \cr
}
}
\author{
Liang He

Maintainer: Liang He <lianghe@mit.edu>
}
\references{
}
\keyword{ package }

\examples{


}
