% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpr2.r
\name{sharpr2}
\alias{sharpr2}
\title{sharpr2}
\usage{
sharpr2(data, l_min = 175, l_max = 450, f_rna = 10, f_dna = 0,
  s_a = 300, verbose = FALSE, auto = TRUE, sig = TRUE, len = FALSE, 
  alpha = 0.05, win = 10, mse = FALSE)
}
\arguments{
\item{data}{A data.frame containing a HiDRA dataset for one chromosome. The data.frame must contain four columns: 'start', 'end', 'PLASMID', 'RNA'}

\item{l_min}{The fragments with a length smaller than l_min will not be processed. The default is 175.}

\item{l_max}{The fragments with a length larger than l_max will not be processed. The default is 450.}

\item{f_rna}{The fragments with an RNA count smaller than f_rna will not be processed. The default is 10.}

\item{f_dna}{The fragments with an DNA count smaller than f_rna will not be processed. The default is 0.}

\item{s_a}{A variance hyperparameter in the prior for the latent regulatory scores. The default is 1000.}

\item{verbose}{An indicator of whether to show processing information. The default is FALSE.}

\item{auto}{An indicator of whether to automatically estimate the ridge coefficient lambda from the data for each tile region. The default is TRUE. If 'auto' is TRUE, 's_a' is ignored and a ridge coefficient is estimated for each tile region separately. If 'auto' is FALSE, a global ridge coefficient (1/s_a) is used.}

\item{sig}{An indicator of whether to identify significant motif regions for the estimated scores. Only valid if auto=TRUE. The default is TRUE.}

\item{len}{An indicator of whether to model log(RNA/PLASMID) of each read as the average or the sum of the latent regulatory scores. The default is FALSE, which is the sum.}

\item{alpha}{A threshold to call the significant region. The default is 0.05.}

\item{win}{A window size for removing sporadic significant regions. If a significant consecutive region is small than win, it will be treated as false signals. The default is 10.}

\item{mse}{An indicator of whether mean square errors are included in the output results. The default is FALSE.}

}
\value{
score: the regulatory scores for each tile region. This list contains four components: est_a (the regulatory scores at each locus), sd_e (the sqare root of the mean square error), var_nb (the variance of the esitmate at each locus), lambda (the ridge coefficient).

region: the start and end positions for each tile region.

n_reg: total number of tile regions.

n_read: the number of reads in each tile region.

sig_reg: identified significant regions based on the cutoff.

motif: predicted 20bp motifs 

cutoff: the cutoff used to call signifnicant regions for the tiled region.
}
\description{
For a HiDRA dataset on a given chromosome, this function calls tile regions (the regions covered by at least one read), and calculates regulatory scores for each tile region. The regulatory scores are based on standardized log(RNA/PLASMID).
}

\details{
The default value of s_a is set to be 300, which is equivalent to a ridge coefficient of 0.0033. This default ridge coefficient value is selected by the median of the estimated lambda from the first library. 

}
\examples{
# sharpr2(data)
}
\keyword{HiDRA}
\keyword{sharpr2}
